Logu
Bitcoin is a modern digital currency developed in 2009 by an anonymous person identified only by Satoshi Nakamoto’s pseudonym. There are no middlemen in transactions – no banks, that is! You can use your Bitcoin for booking Expedia hotels, shopping for furniture, and many more. But a great deal of the hype is to make it valuable. In 2017, the pr...